Refactor ScrambleRequests bundling: ZIP, PDF
