+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/pdf/PdfContent.kt
@@ -1,0 +1,5 @@
+package org.worldcubeassociation.tnoodle.server.webscrambles.pdf
+
+interface PdfContent {
+    fun render(): ByteArray
+}
