+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/ScrambleHandler.kt
@@ -67,14 +67,14 @@ object ScrambleHandler : RouteHandler {
                 }
                 "json" -> call.respond(scrambleRequests)
                 "pdf" -> {
-                    val totalPdfOutput = ScrambleRequest.requestsToPdf(title, generationDate, scrambleRequests, null)
+                    val totalPdfOutput = ScrambleRequest.requestsToCompletePdf(title, generationDate, scrambleRequests, null)
                     call.response.header("Content-Disposition", "inline")
 
                     // Workaround for Chrome bug with saving PDFs:
                     //  https://bugs.chromium.org/p/chromium/issues/detail?id=69677#c35
                     call.response.header("Cache-Control", "public")
 
-                    call.respondBytes(totalPdfOutput.toByteArray(), ContentType.Application.Pdf)
+                    call.respondBytes(totalPdfOutput.render(), ContentType.Application.Pdf)
                 }
                 "zip" -> {
                     val baosZip = ScrambleRequest.requestsToZip(title, generationDate, scrambleRequests, seed, generationUrl, null)
